pyproject.toml
[build-system]
requires = ["setuptools >= 61.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pytest-balance-xdist"
description = "Divide tests into evenly balanced workloads for each xdist worker."
authors = [
    {name = "Richard Si", email = "sichard26@gmail.com"}
]
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Framework :: Pytest",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Testing",
]
license = {file = "LICENSE.txt"}
dependencies = ["pytest >= 6.0.0", "pytest-xdist >= 1.34.0"]
requires-python = ">=3.7"
dynamic = ["version"]

[project.urls]
GitHub = "https://github.com/ichard26/pytest-balance-xdist"

[tool.setuptools]
dynamic = { version = { attr = "pytest_balance_xdist.__version__" } }

[project.scripts]
balance-xdist-show-worker-times = "pytest_balance_xdist:cli_entrypoint"

[project.entry-points]
pytest11 = { balance-xdist = "pytest_balance_xdist" }

[tool.black]
line-length = 95

[tool.isort]
profile = "black"
known_third_party = ["pytest", "xdist"]
